+++ ospray/lights/DirectionalLight.cpp
@@ -3,12 +3,17 @@
 
 #include "DirectionalLight.h"
 #include "lights/DirectionalLight_ispc.h"
+#include "lights/Light_ispc.h"
 
 namespace ospray {
 
-DirectionalLight::DirectionalLight()
+void *DirectionalLight::createIE(const void *instance) const
 {
-  ispcEquivalent = ispc::DirectionalLight_create();
+  void *ie = ispc::DirectionalLight_create();
+  ispc::Light_set(ie, visible, (const ispc::Instance *)instance);
+  ispc::DirectionalLight_set(
+      ie, (ispc::vec3f &)irradiance, (ispc::vec3f &)direction, cosAngle);
+  return ie;
 }
 
 std::string DirectionalLight::toString() const
@@ -25,16 +30,13 @@ void DirectionalLight::commit()
   // the ispc::DirLight expects direction towards light source
   direction = -normalize(direction);
   angularDiameter = clamp(angularDiameter, 0.f, 180.f);
-  const float cosAngle = std::cos(deg2rad(0.5f * angularDiameter));
+  cosAngle = std::cos(deg2rad(0.5f * angularDiameter));
 
   queryIntensityQuantityType(OSP_INTENSITY_QUANTITY_IRRADIANCE);
-  processIntensityQuantityType(cosAngle);
-
-  ispc::DirectionalLight_set(
-      getIE(), (ispc::vec3f &)irradiance, (ispc::vec3f &)direction, cosAngle);
+  processIntensityQuantityType();
 }
 
-void DirectionalLight::processIntensityQuantityType(const float &cosAngle)
+void DirectionalLight::processIntensityQuantityType()
 {
   // validate the correctness of the light quantity type
   if (intensityQuantity == OSP_INTENSITY_QUANTITY_RADIANCE) {